*** empty log message ***
git-svn-id: https://ceph.svn.sf.net/svnroot/ceph@825 29311d96-e01e-0410-9327-a35deaab8ce9

+++ ceph/osd/ObjectStore.h
@@ -47,7 +47,7 @@ public:
 	int avg_extent;
 	map<int,int> extent_dist;          // powers of two
 
-	int avg_extent_per_object;
+	float avg_extent_per_object;
 	int avg_extent_jump;  // avg distance bweteen consecutive extents
 
 	int num_free_extent;
